+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "aria-docs",
+  "name": "@rocicorp/zero-docs",
   "version": "1.0.0",
   "private": true,
   "scripts": {
